+++ opentelemetry-etw-logs/src/logs/converters.rs
@@ -17,6 +17,11 @@ impl IntoJson for AnyValue {
             AnyValue::Bytes(_value) => todo!("No support for AnyValue::Bytes yet."),
             AnyValue::ListAny(value) => value.as_json_value(),
             AnyValue::Map(value) => value.as_json_value(),
+            &&_ => {
+                // If we reach this line is because we are upgrading opentelemetry to a version that contains a new type.
+                // We must add support for the new type.
+                todo!("Add support for new types");
+            }
         }
     }
 }