+++ opentelemetry-etw-logs/src/logs/reentrant_logprocessor.rs
@@ -1,10 +1,10 @@
 use std::fmt::Debug;
 
-use opentelemetry::logs::LogResult;
-use opentelemetry::InstrumentationLibrary;
+use opentelemetry::InstrumentationScope;
 use opentelemetry_sdk::logs::LogRecord;
+use opentelemetry_sdk::logs::LogResult;
 
-#[cfg(feature = "logs_level_enabled")]
+#[cfg(feature = "spec_unstable_logs_enabled")]
 use opentelemetry_sdk::export::logs::LogExporter;
 
 use crate::logs::exporter::ExporterConfig;
@@ -33,7 +33,7 @@ impl ReentrantLogProcessor {
 }
 
 impl opentelemetry_sdk::logs::LogProcessor for ReentrantLogProcessor {
-    fn emit(&self, data: &mut LogRecord, instrumentation: &InstrumentationLibrary) {
+    fn emit(&self, data: &mut LogRecord, instrumentation: &InstrumentationScope) {
         _ = self.event_exporter.export_log_data(data, instrumentation);
     }
 
@@ -49,7 +49,7 @@ impl opentelemetry_sdk::logs::LogProcessor for ReentrantLogProcessor {
         Ok(())
     }
 
-    #[cfg(feature = "logs_level_enabled")]
+    #[cfg(feature = "spec_unstable_logs_enabled")]
     fn event_enabled(
         &self,
         level: opentelemetry::logs::Severity,